requirements.txt
faker
pandas
boto3
python-dotenv
numpy
openpyxl      
unidecode
pyarrow
joblib
scikit-learn